# specs/001-audit-and-enforce/contracts/predict.openapi.yaml
openapi: 3.1.0
info:
  title: Retail Demand Forecasting API
  version: 1.0.0
  description: Tweedie LightGBM for zero-inflated demand, hierarchical store/dept/cat trends (Principle I, IV)
paths:
  /predict:
    post:
      summary: Single forecast (float preserved)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictionRequest'
      responses:
        '200':
          description: float prediction (e.g., 0.15 for HOBBIES_1_003)
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction: { type: number, minimum: 0 }
                  features: { type: object }
        '422': { description: Pydantic validation error (extra="forbid", ge/le bounds) }
        '503': { description: Model not loaded }
  /predict/batch:
    post:
      summary: Batch up to 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchPredictionRequest'
      responses:
        '200':
          description: batch predictions
        '422': { description: batch size 0 or >1000 or validation }
components:
  schemas:
    PredictionRequest:
      type: object
      additionalProperties: false
      required: [lag_1, lag_2, lag_3, lag_7, lag_14, lag_21, lag_28, rolling_mean_7, rolling_min_7, rolling_max_7, rolling_std_7, rolling_mean_28, rolling_min_28, rolling_max_28, rolling_std_28, store_rolling_mean_7, store_rolling_mean_28, dept_rolling_mean_7, dept_rolling_mean_28, cat_rolling_mean_7, cat_rolling_mean_28, day_of_week, day_of_month, month, year, is_weekend, day_of_week_sin, day_of_week_cos, day_of_month_sin, day_of_month_cos, month_sin, month_cos, snap_CA, snap_TX, snap_WI, has_event_1, has_event_2, sell_price]
      properties:
        lag_1: { type: number, minimum: 0, maximum: 10000, description: Sales 1 day ago (float preserved) }
        lag_2: { type: number, minimum: 0, maximum: 10000 }
        lag_3: { type: number, minimum: 0, maximum: 10000 }
        lag_7: { type: number, minimum: 0, maximum: 10000 }
        lag_14: { type: number, minimum: 0, maximum: 10000 }
        lag_21: { type: number, minimum: 0, maximum: 10000 }
        lag_28: { type: number, minimum: 0, maximum: 10000 }
        rolling_mean_7: { type: number, minimum: 0, maximum: 10000 }
        rolling_min_7: { type: number, minimum: 0, maximum: 10000 }
        rolling_max_7: { type: number, minimum: 0, maximum: 10000 }
        rolling_std_7: { type: number, minimum: 0, maximum: 10000 }
        rolling_mean_28: { type: number, minimum: 0, maximum: 10000 }
        rolling_min_28: { type: number, minimum: 0, maximum: 10000 }
        rolling_max_28: { type: number, minimum: 0, maximum: 10000 }
        rolling_std_28: { type: number, minimum: 0, maximum: 10000 }
        store_rolling_mean_7: { type: number, minimum: 0, maximum: 10000 }
        store_rolling_mean_28: { type: number, minimum: 0, maximum: 10000 }
        dept_rolling_mean_7: { type: number, minimum: 0, maximum: 10000 }
        dept_rolling_mean_28: { type: number, minimum: 0, maximum: 10000 }
        cat_rolling_mean_7: { type: number, minimum: 0, maximum: 10000 }
        cat_rolling_mean_28: { type: number, minimum: 0, maximum: 10000 }
        day_of_week: { type: integer, minimum: 1, maximum: 7 }
        day_of_month: { type: integer, minimum: 1, maximum: 31 }
        month: { type: integer, minimum: 1, maximum: 12 }
        year: { type: integer, minimum: 2000, maximum: 2030 }
        is_weekend: { type: integer, minimum: 0, maximum: 1 }
        day_of_week_sin: { type: number, minimum: -1, maximum: 1 }
        day_of_week_cos: { type: number, minimum: -1, maximum: 1 }
        day_of_month_sin: { type: number, minimum: -1, maximum: 1 }
        day_of_month_cos: { type: number, minimum: -1, maximum: 1 }
        month_sin: { type: number, minimum: -1, maximum: 1 }
        month_cos: { type: number, minimum: -1, maximum: 1 }
        snap_CA: { type: integer, minimum: 0, maximum: 1 }
        snap_TX: { type: integer, minimum: 0, maximum: 1 }
        snap_WI: { type: integer, minimum: 0, maximum: 1 }
        has_event_1: { type: integer, minimum: 0, maximum: 1 }
        has_event_2: { type: integer, minimum: 0, maximum: 1 }
        sell_price: { type: number, minimum: 0, maximum: 10000 }
    BatchPredictionRequest:
      type: object
      properties:
        requests:
          type: array
          minItems: 1
          maxItems: 1000
          items: { $ref: '#/components/schemas/PredictionRequest' }